Chat notifications.File attachments,Company data,Profile image management,
