+++ src/backend/index.js
@@ -37,6 +37,11 @@ import commentRoutes from './routes/commentRoutes.js';
 import ticketUpdateRoutes from './routes/ticketUpdateRoutes.js';
 import dashboardRoutes from './routes/dashboardRoutes.js';
 import userDashboardRoutes from './routes/userDashboardRoutes.js';
+import attachmentRoutes from './routes/attachmentRoutes.js';
+import ticketManagementRoutes from './routes/ticketManagementRoutes.js';
+import companyRoutes from './routes/companyRoutes.js';
+import profileRoutes from './routes/profileRoutes.js';
+import chatNotificationRoutes from './routes/chatNotificationRoutes.js';
 
 const app = express();
 app.use(bodyParser.json());
@@ -77,6 +82,11 @@ app.use('/api', authLoginRoutes);
 app.use('/api', authPasswordRoutes);
 app.use('/api', authResetRoutes);
 app.use('/api/user/tickets', userDashboardRoutes);
+app.use('/api', attachmentRoutes);
+app.use('/api', ticketManagementRoutes);
+app.use('/api', companyRoutes);
+app.use('/api', profileRoutes);
+app.use('/api', chatNotificationRoutes);
 
 //evidence uploads
 app.use("/uploads", express.static("uploads"));
@@ -1468,7 +1478,7 @@ app.post('/api/upload_evidence', upload_evidence.array('evidenceFiles'), async (
   }
 });
 
-
+//x
 
 // Add ticket log routes
 app.use('/api/ticket-logs', ticketLogRoutes);
@@ -1480,485 +1490,3 @@ server.listen(PORT, () => {
 });
 
 export default db;
-
-// Add attachment endpoint
-app.post('/api/tickets/:ticketId/attachments', async (req, res) => {
-  const { ticketId } = req.params;
-  const { fileName, fileUrl, userId } = req.body;
-
-  try {
-    // Create log entry for the attachment
-    const logResult = await createTicketLog(
-      ticketId,
-      'ATTACHMENT',
-      `added an attachment: ${fileName}`,
-      userId,
-      null,
-      fileUrl
-    );
-
-    // Get ticket creator ID for notification
-    const getTicket = "SELECT UserId as ticketUserId FROM ticket WHERE TicketID = ?";
-    const [ticketResult] = await new Promise((resolve, reject) => {
-      db.query(getTicket, [ticketId], (err, results) => {
-        if (err) reject(err);
-        else resolve(results);
-      });
-    });
-
-    if (ticketResult) {
-      try {
-        await createNotification(
-          ticketResult.ticketUserId,
-          `New attachment added to your ticket #${ticketId}`,
-          'NEW_ATTACHMENT',
-          logResult.insertId
-        );
-      } catch (error) {
-        console.error("Error creating notification:", error);
-      }
-    }
-
-    res.json({ message: "Attachment added successfully" });
-  } catch (error) {
-    console.error("Error adding attachment:", error);
-    res.status(500).json({ message: "Server error" });
-  }
-});
-
-// Configure multer for attachments
-const attachmentStorage = multer.diskStorage({
-  destination: (req, file, cb) => {
-    const uploadDir = path.join(__dirname, 'uploads', 'attachments');
-    fs.mkdirSync(uploadDir, { recursive: true });
-    cb(null, uploadDir);
-  },
-  filename: (req, file, cb) => {
-    const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1E9);
-    cb(null, uniqueSuffix + '-' + file.originalname);
-  }
-});
-
-const attachmentUpload = multer({
-  storage: attachmentStorage,
-  limits: {
-    fileSize: 10 * 1024 * 1024 // 10MB limit
-  }
-});
-
-// Add attachment endpoint
-app.post('/api/tickets/:ticketId/attachments', attachmentUpload.single('file'), async (req, res) => {
-  const { ticketId } = req.params;
-  const { userId } = req.body;
-  const file = req.file;
-
-  if (!file) {
-    return res.status(400).json({ message: "No file uploaded" });
-  }
-
-  // Get user details
-  const getUserQuery = `
-        SELECT 
-            au.FullName as uploaderName,
-            t.UserId as ticketUserId,
-            t.SupervisorID
-        FROM appuser au
-        LEFT JOIN ticket t ON t.TicketID = ?
-        WHERE au.UserID = ?
-    `;
-
-  try {
-    const [userResults] = await new Promise((resolve, reject) => {
-      db.query(getUserQuery, [ticketId, userId], (err, results) => {
-        if (err) reject(err);
-        else resolve(results);
-      });
-    });
-
-    if (!userResults) {
-      return res.status(404).json({ message: "User or ticket not found" });
-    }
-
-    const uploaderName = userResults.uploaderName;
-    const filePath = file.path;
-    const fileName = file.originalname;
-    const fileSize = file.size;
-    const fileType = file.mimetype;
-
-    // Create ticket log entry
-    const logQuery = `
-            INSERT INTO ticketlog 
-            (TicketID, DateTime, Type, Description, UserID, Note, NewValue)
-            VALUES (?, NOW(), ?, ?, ?, ?, ?)
-        `;
-
-    const description = `uploaded file "${fileName}"`;
-    const fileDetails = JSON.stringify({
-      name: fileName,
-      size: fileSize,
-      type: fileType,
-      path: filePath
-    });
-
-    const [logResult] = await createTicketLog(
-      ticketId,
-      'ATTACHMENT',
-      description,
-      userId,
-      null,
-      fileDetails,
-      `File size: ${Math.round(fileSize / 1024)}KB`
-    );
-
-    // Update LastRespondedTime
-    await new Promise((resolve, reject) => {
-      db.query(
-        "UPDATE ticket SET LastRespondedTime = NOW() WHERE TicketID = ?",
-        [ticketId],
-        (err, result) => {
-          if (err) reject(err);
-          else resolve(result);
-        }
-      );
-    });
-
-    // Create notifications
-    try {
-      // Notify ticket creator if attachment is from someone else
-      if (userId !== userResults.ticketUserId) {
-        await createNotification(
-          userResults.ticketUserId,
-          `New file "${fileName}" attached to ticket #${ticketId} by ${uploaderName}`,
-          'NEW_ATTACHMENT',
-          logResult.insertId
-        );
-      }
-
-      // Notify supervisor if exists and attachment is from ticket creator
-      if (userResults.SupervisorID && userId === userResults.ticketUserId) {
-        await createNotification(
-          userResults.SupervisorID,
-          `Ticket creator attached a file "${fileName}" to ticket #${ticketId}`,
-          'CREATOR_ATTACHMENT',
-          logResult.insertId
-        );
-      }
-    } catch (error) {
-      console.error("Error creating notifications:", error);
-    }
-
-    res.json({
-      message: "File uploaded successfully",
-      file: {
-        name: fileName,
-        path: filePath,
-        size: fileSize,
-        type: fileType
-      },
-      logId: logResult.insertId
-    });
-  } catch (error) {
-    console.error("Error uploading file:", error);
-    res.status(500).json({ message: "Server error" });
-  }
-});
-
-// Assign supervisor to ticket
-app.put('/api/tickets/:ticketId/supervisor', async (req, res) => {
-  const { ticketId } = req.params;
-  const { supervisorId, userId } = req.body;
-
-  // Get current ticket and user details
-  const getDetailsQuery = `
-        SELECT 
-            t.SupervisorID as currentSupervisorId,
-            t.UserId as ticketUserId,
-            au_new.FullName as newSupervisorName,
-            au_old.FullName as currentSupervisorName,
-            au_assigner.FullName as assignerName
-        FROM ticket t
-        LEFT JOIN appuser au_new ON au_new.UserID = ?
-        LEFT JOIN appuser au_old ON au_old.UserID = t.SupervisorID
-        LEFT JOIN appuser au_assigner ON au_assigner.UserID = ?
-        WHERE t.TicketID = ?
-    `;
-
-  try {
-    const [details] = await new Promise((resolve, reject) => {
-      db.query(getDetailsQuery, [supervisorId, userId, ticketId], (err, results) => {
-        if (err) reject(err);
-        else resolve(results);
-      });
-    });
-
-    if (!details) {
-      return res.status(404).json({ message: "Ticket not found" });
-    }
-
-    // Update supervisor
-    const updateQuery = "UPDATE ticket SET SupervisorID = ?, LastRespondedTime = NOW() WHERE TicketID = ?";
-
-    await new Promise((resolve, reject) => {
-      db.query(updateQuery, [supervisorId, ticketId], (err, result) => {
-        if (err) reject(err);
-        else resolve(result);
-      });
-    });
-
-    // Create ticket log entry
-    const logQuery = `
-            INSERT INTO ticketlog 
-            (TicketID, DateTime, Type, Description, UserID, OldValue, NewValue, Note)
-            VALUES (?, NOW(), ?, ?, ?, ?, ?, ?)
-        `;
-
-    let description;
-    if (!details.currentSupervisorId) {
-      description = `Supervisor assigned: ${details.newSupervisorName}`;
-    } else {
-      description = `Supervisor changed from ${details.currentSupervisorName} to ${details.newSupervisorName}`;
-    }
-
-    const note = `Updated by ${details.assignerName}`;
-
-    const [logResult] = await new Promise((resolve, reject) => {
-      db.query(
-        logQuery,
-        [
-          ticketId,
-          'SUPERVISOR_CHANGE',
-          description,
-          userId,
-          details.currentSupervisorId,
-          supervisorId,
-          note
-        ],
-        (err, result) => {
-          if (err) reject(err);
-          else resolve(result);
-        }
-      );
-    });
-
-    // Create notifications
-    try {
-      // Notify new supervisor
-      await createNotification(
-        supervisorId,
-        `You have been assigned as supervisor for ticket #${ticketId} by ${details.assignerName}`,
-        'SUPERVISOR_ASSIGNED',
-        logResult.insertId
-      );
-
-      // Notify ticket creator
-      await createNotification(
-        details.ticketUserId,
-        `${details.newSupervisorName} has been assigned as supervisor for your ticket #${ticketId}`,
-        'SUPERVISOR_UPDATED',
-        logResult.insertId
-      );
-
-      // Notify old supervisor if exists
-      if (details.currentSupervisorId && details.currentSupervisorId !== supervisorId) {
-        await createNotification(
-          details.currentSupervisorId,
-          `You have been unassigned from ticket #${ticketId} by ${details.assignerName}`,
-          'SUPERVISOR_UNASSIGNED',
-          logResult.insertId
-        );
-      }
-    } catch (error) {
-      console.error("Error creating notifications:", error);
-    }
-
-    res.json({
-      message: "Supervisor assigned successfully",
-      logId: logResult.insertId
-    });
-  } catch (error) {
-    console.error("Error assigning supervisor:", error);
-    res.status(500).json({ message: "Server error" });
-  }
-});
-
-// Update ticket due date
-app.put('/api/tickets/:ticketId/due-date', async (req, res) => {
-  const { ticketId } = req.params;
-  const { dueDate, userId } = req.body; // <-- userId must be sent from frontend
-
-  db.query('SELECT DueDate FROM ticket WHERE TicketID = ?', [ticketId], (err, results) => {
-    if (err || results.length === 0) return res.status(500).json({ message: 'Error' });
-    const oldDueDate = results[0].DueDate;
-
-    db.query('UPDATE ticket SET DueDate = ? WHERE TicketID = ?', [dueDate, ticketId], async (err2) => {
-      if (err2) return res.status(500).json({ message: 'Error updating due date' });
-
-      let desc;
-      if (!oldDueDate) {
-        desc = `changed due date to ${dueDate}`;
-      } else {
-        desc = `changed due date from ${oldDueDate} to ${dueDate}`;
-      }
-      await createTicketLog(
-        ticketId,
-        'DUE_DATE_CHANGE',
-        desc,
-        userId,
-        oldDueDate,
-        dueDate,
-        null
-      );
-      res.json({ message: 'Due date updated and logged' });
-    });
-  });
-});
-
-// Update ticket resolution
-app.put('/api/tickets/:ticketId/resolution', async (req, res) => {
-  const { ticketId } = req.params;
-  const { resolution, userId } = req.body; // <-- userId must be sent from frontend
-
-  db.query('SELECT Resolution FROM ticket WHERE TicketID = ?', [ticketId], (err, results) => {
-    if (err || results.length === 0) return res.status(500).json({ message: 'Error' });
-    const oldResolution = results[0].Resolution;
-
-    db.query('UPDATE ticket SET Resolution = ? WHERE TicketID = ?', [resolution, ticketId], async (err2) => {
-      if (err2) return res.status(500).json({ message: 'Error updating resolution' });
-
-      let desc;
-      if (!oldResolution) {
-        desc = `changed resolution to "${resolution || ''}"`;
-      } else {
-        desc = `changed resolution from "${oldResolution || ''}" to "${resolution || ''}"`;
-      }
-      await createTicketLog(
-        ticketId,
-        'RESOLUTION_CHANGE',
-        desc,
-        userId,
-        oldResolution,
-        resolution,
-        null
-      );
-      res.json({ message: 'Resolution updated and logged' });
-    });
-  });
-});
-
-
-// API endpoint to fetch companies
-app.get('/api/companies', (req, res) => {
-  const query = `
-    SELECT DISTINCT c.CompanyName
-    FROM client c
-    JOIN appuser au ON c.UserID = au.UserID
-    WHERE c.CompanyName IS NOT NULL
-    ORDER BY c.CompanyName
-  `;
-
-  db.query(query, (err, results) => {
-    if (err) {
-      console.error('Error fetching companies:', err);
-      return res.status(500).json({ error: 'Failed to fetch companies' });
-    }
-    res.json(results);
-  });
-});
-
-// Add Multer middleware for profile image uploads
-app.post(
-  "/api/user/profile/upload/:id",
-  upload.single("profileImage"),
-  async (req, res) => {
-    const userId = req.params.id;
-    if (!req.file) {
-      return res.status(400).json({ message: "No file uploaded." });
-    }
-
-    const imagePath = `profile_images/${req.file.filename}`; // Store relative path
-
-    try {
-      // Update the user's profile image path in the database
-      const updateQuery =
-        "UPDATE appuser SET ProfileImagePath = ? WHERE UserID = ?";
-      await db.promise().query(updateQuery, [imagePath, userId]);
-
-      res.status(200).json({
-        message: "Profile image uploaded successfully",
-        imagePath: imagePath, // Send back the relative path
-      });
-    } catch (error) {
-      console.error("Error updating profile image path in DB:", error);
-      res.status(500).json({ message: "Failed to update profile image." });
-    }
-  }
-);
-
-app.delete('/api/user/profile/image/:id', async (req, res) => {
-  const userId = req.params.id;
-  try {
-    // Get the current image path to delete the file from the server
-    const [results] = await db.promise().query('SELECT ProfileImagePath FROM appuser WHERE UserID = ?', [userId]);
-    if (results.length > 0 && results[0].ProfileImagePath) {
-      const filePath = path.join(__dirname, 'uploads', results[0].ProfileImagePath);
-      fs.unlink(filePath, (err) => {
-        if (err) {
-          console.error('Error deleting old profile image file:', err);
-          // Continue even if file deletion fails, as DB update is more critical
-        }
-      });
-    }
-
-    // Update the database to remove the profile image path
-    const updateQuery = 'UPDATE appuser SET ProfileImagePath = NULL WHERE UserID = ?';
-    await db.promise().query(updateQuery, [userId]);
-
-    res.status(200).json({ message: 'Profile image removed successfully' });
-  } catch (error) {
-    console.error('Error removing profile image:', error);
-    res.status(500).json({ message: 'Failed to remove profile image.' });
-  }
-});
-
-// API endpoint to get unread chat notifications count for a specific ticket
-app.get('/api/notifications/chat/count/:userId/:ticketId', (req, res) => {
-  const { userId, ticketId } = req.params;
-  const query = `
-        SELECT COUNT(*) as count
-        FROM notifications
-        WHERE UserID = ?
-        AND TicketLogID = ?  /* For chat messages, TicketLogID stores TicketID */
-        AND Type = 'NEW_CHAT_MESSAGE'
-        AND IsRead = FALSE;
-    `;
-
-  db.query(query, [userId, ticketId], (err, results) => {
-    if (err) {
-      console.error('Error fetching unread chat notification count:', err);
-      return res.status(500).json({ error: 'Failed to fetch unread chat notification count' });
-    }
-    res.json({ count: results[0].count });
-  });
-});
-
-// API endpoint to mark chat notifications as read for a specific ticket and user
-app.put('/api/notifications/chat/read/:userId/:ticketId', (req, res) => {
-  const { userId, ticketId } = req.params;
-
-  const query = `
-        UPDATE notifications
-        SET IsRead = TRUE
-        WHERE UserID = ?
-        AND TicketLogID = ?  /* For chat messages, TicketLogID stores TicketID */
-        AND Type = 'NEW_CHAT_MESSAGE'
-        AND IsRead = FALSE;
-    `;
-
-  db.query(query, [userId, ticketId], (err, result) => {
-    if (err) {
-      console.error('Error marking chat notifications as read:', err);
-      return res.status(500).json({ error: 'Failed to mark chat notifications as read' });
-    }
-    res.json({ message: 'Chat notifications marked as read', updatedCount: result.affectedRows });
-  });
-});
